refactor: DefaultServlet 제거 및 ResourceReader util로 변경
