+++ tomcat/src/main/java/org/apache/catalina/controller/IndexController.java
@@ -1,5 +1,7 @@
 package org.apache.catalina.controller;
 
+import java.io.IOException;
+import org.apache.catalina.util.ResourceReader;
 import org.apache.coyote.http11.HttpMethod;
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
@@ -7,7 +9,7 @@ import org.apache.coyote.http11.HttpResponse;
 public class IndexController implements Controller {
 
     @Override
-    public void service(HttpRequest request, HttpResponse response) {
+    public void service(HttpRequest request, HttpResponse response) throws IOException {
         HttpMethod method = request.getMethod();
 
         if (method.isGet()) {
@@ -15,7 +17,7 @@ public class IndexController implements Controller {
         }
     }
 
-    public void doGet(HttpResponse response) {
-        response.setRedirect("/index.html");
+    public void doGet(HttpResponse response) throws IOException {
+        ResourceReader.serveResource("/index.html", response);
     }
 }
